==> Notebooks/03_Analisis_Sentimientos.ipynb <==
# 03.Analisis de Sentimientos

## Importacion de los datos

import pandas as pd

df_news=pd.read_csv('../Datos/noticias_bitcoin_top3_diarias.csv')

df_news.info()

## Instalcion librerias para procesar sentimientos

!pip install transformers torch

df_news['contenido'][0]

## Entrenamiento del modelo para analisis de sentimientos

from transformers import AutoTokenizer, AutoModelForSequenceClassification
import torch
import torch.nn.functional as F
import pandas as pd

# Cargar modelo y tokenizer
model_name = 'cardiffnlp/twitter-roberta-base-sentiment-latest'
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSequenceClassification.from_pretrained(model_name)

# Función para calcular el sentimiento de una frase
def get_sentiment_score(text):
    inputs = tokenizer(text, return_tensors="pt", truncation=True)
    outputs = model(**inputs)
    probs = F.softmax(outputs.logits, dim=1).detach().numpy()[0]
    neg, neu, pos = probs
    score = -1 * neg + 0 * neu + 1 * pos
    return score

# Aplicar a la columna de títulos
df_news['score'] = df_news['titulo'].apply(get_sentiment_score)

# Mostrar resultados
print(df_news[['titulo', 'score']].head())


### Visualizacion de metricas de sentimientos

print(df_news['titulo'][100])
print(df_news['score'][100])

## Exportacion del Archivo con sentimientos

pd.DataFrame(df_news).to_csv('../Datos/noticias_bitcoin_sentimientos.csv', index=False)
